// src/enums/admin/plans.ts
export enum PlanDetailFields {
  name = "planName",
  priceMonthly = "monthlyPrice",
  anuallDiscount = "discount",
  employeeNo = "numberOfEmployees",
  requestNo = "numberOfRequests",
  description = "description",
  accountingReports = "accountingReport",
  customEmails = "customizedEmail",
  waterMark = "watermark",
  apifeature = "apiFeatures",
}